==> xLTS/06_xing_attr.ipynb <==
import geopandas as gpd
import os
import warnings
import ast

warnings.filterwarnings("ignore")

# Base directory path
base_dir = r"C:\Users\natda\OneDrive - Northeastern University\Desktop\NatDave\Academics\PhD_NU\RESEARCH\Traffic_Stress\Boston"

# Construct file paths
roads_path = os.path.join(base_dir, "street_network.shp")
crossings_path = os.path.join(base_dir, "crossings.shp")

# Load shapefiles
roads = gpd.read_file(roads_path)
crossings = gpd.read_file(crossings_path)

# Convert FRM_LEG, TO_LEG, and CRS_LEG from string to list
crossings['FRM_LEG'] = crossings['FRM_LEG'].apply(ast.literal_eval)
crossings['TO_LEG'] = crossings['TO_LEG'].apply(ast.literal_eval)
crossings['CRS_LEG'] = crossings['CRS_LEG'].apply(ast.literal_eval)

# Create mapping dictionaries from the street network
bike_type_dict = roads.set_index('unique_id')['bike_type2'].to_dict()
stoperneu_dict = roads.set_index('unique_id')['StOperNEU'].to_dict()
federalfun_dict = roads.set_index('unique_id')['FEDERALFUN'].to_dict()
ADT_2025_dict = roads.set_index('unique_id')['ADT_2025'].to_dict()
numlanes_dict = roads.set_index('unique_id')['NUMLANE'].to_dict()
speed_dict = roads.set_index('unique_id')['SPEED'].to_dict()
qNoAccess_dict = roads.set_index('unique_id')['qNoAccess'].to_dict()

# Function to get first element value from mapping dictionary
def get_first_leg_value(leg_list, mapping_dict):
    if leg_list:
        return mapping_dict.get(leg_list[0], "")  # Leave empty if no match
    return ""

# Assign values based on first element in FRM_LEG, TO_LEG, and CRS_LEG
crossings['FR_BK_TYPE'] = crossings['FRM_LEG'].apply(lambda x: get_first_leg_value(x, bike_type_dict))
crossings['TO_BK_TYPE'] = crossings['TO_LEG'].apply(lambda x: get_first_leg_value(x, bike_type_dict))

crossings['FR_StOpNU'] = crossings['FRM_LEG'].apply(lambda x: get_first_leg_value(x, stoperneu_dict))
crossings['TO_StOpNU'] = crossings['TO_LEG'].apply(lambda x: get_first_leg_value(x, stoperneu_dict))

crossings['FR_CLASS'] = crossings['FRM_LEG'].apply(lambda x: get_first_leg_value(x, federalfun_dict))
crossings['TO_CLASS'] = crossings['TO_LEG'].apply(lambda x: get_first_leg_value(x, federalfun_dict))

crossings['CRS_ADT'] = crossings['CRS_LEG'].apply(lambda x: get_first_leg_value(x, ADT_2025_dict))

crossings['CRS_LANES'] = crossings['CRS_LEG'].apply(lambda x: get_first_leg_value(x, numlanes_dict))

crossings['CRS_SPEED'] = crossings['CRS_LEG'].apply(lambda x: get_first_leg_value(x, speed_dict))

crossings['CRS_qNoAcc'] = crossings['CRS_LEG'].apply(lambda x: get_first_leg_value(x, qNoAccess_dict))

# Save the updated crossings shapefile
crossings.to_file(crossings_path, driver='ESRI Shapefile')

print("Updated crossings.shp with speed limit, bike type, StOperNEU, functional classifications, ADT, qNoAccess and number of lanes.")
print("CROSS_DIST column initialized for future population.")

print(roads['ADT_2025'].value_counts(dropna=False))
